Remove go.sum from project
Go modules support is still experimental in 1.11.
Even though the upstream recommendation is to check-in the go.sum file,
we decided to remove it until we do a release or the modules feature
becomes stable.
This way we can avoid checksum mismatches on our CI due to different Go
versions or upstream changes.

See:
https://github.com/golang/go/wiki/Modules#is-gosum-a-lock-file-why-does-gosum-include-information-for-module-versions-i-am-no-longer-using