Changing delete event such that duplicate events are removed
Multiple delete events were being thrown in the trigger reconciler
code. Ideally, there should be a single delete event. This commit is
defining the delete event as the one, when grace period is set to
zero in pod yaml because that is when the pod is issued KILL signal
and removed from the kube database.

Ref :- https://kubernetes.io/docs/concepts/workloads/pods/pod/#termination-of-pods