Fix flaky test case of output persistor
The directory in which the flaky test case wrote the `provides.json` is
also used by other test cases and the directory being used is also
required to be present due to other test cases.

Create a temporary directory just for the test case and configure the
output persitor to use the temporary directory instead of `/tmp`.